unify the paging mechanism and caching mechanisms of indexes and move them to core project add paging mechanisms to the JDBC index implementations and the SQL query processor

+++ majortom-api/src/main/java/de/topicmapslab/majortom/model/core/paged/IPagedName.java
@@ -63,6 +63,6 @@ public interface IPagedName {
 	 * @throws UnsupportedOperationException
 	 *             thrown if paging is not supported by the underlying store
 	 */
-	public int getNumberOfVariants();
+	public long getNumberOfVariants();
 
 }
